Added testing for the Collectable Database and the Collectable.
Added default values to the Collectable to prevent returning null values.
Added a deleteCollectable, exists, and get function to the database to directly delete, check the existence of, and directly return a collectable within the database.
Allowed the callback on the CollectableDatabase::add member to be nulled, allowing it to be used outside of the HomeActivity.