QR Codes can now be scanned by more than one Player. Player metrics are calculated at runtime rather than through a stored value.
Because QR Codes are stored separately from the player, this means that when a player scans a code, they will then be able to overwrite the Name/Geolocation/Photo for all users who have scanned this code. This is kinda bad, but because we didn't setup our system this way, we can't really do much unless we want to overhaul the entire system.

Player metrics being uploaded as a raw value led to them getting out of sync, and not being updated. While it may technically be slower, calculating each metric on demand ensures that it's always up to date.